# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Score API",
        "contact": {
            "name": "Score API Developer Team",
            "url": "https://github.com/pine09/ScoreAPI"
        },
        "version": "1.0"
    },
    "host": "score.app",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/admin/jurusan": {
            "get": {
                "tags": [
                    "Admin on Jurusan"
                ],
                "summary": "Mengambil semua jurusan yang ada.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data jurusan. (max 5 per page)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jurusan"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin on Jurusan"
                ],
                "summary": "Mendaftarkan jurusan baru.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data jurusan Baru.",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jurusan"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data jurusan baru."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "409": {
                        "description": "Conflict."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    }
                }
            }
        },
        "/api/v1/admin/jurusan/{id}": {
            "get": {
                "tags": [
                    "Admin on Jurusan"
                ],
                "summary": "Mengambil data jurusan spesifik",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data jurusan yang dicari"
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin on Jurusan"
                ],
                "summary": "Mengedit data jurusan.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data jurusan (Edit).",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jurusan"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data jurusan edited."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    },
                    "409": {
                        "description": "Conflict."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin on Jurusan"
                ],
                "summary": "Men-delete data jurusan.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data jurusan deleted."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jurusan": {
            "properties": {
                "jurusan_code": {
                    "type": "string"
                },
                "jurusan_name": {
                    "type": "string"
                }
            }
        },
...
    }
}